Feature: features/도메인/model로직 page에 적용
