+++ src/features/comments/comment-list/index.ts
@@ -1,0 +1,2 @@
+export * from "./ui/CommentList"
+export * from "./model/useCommentList"
